Refactoring module loading
