+++ packages/mux-uploader/src/constants.ts
@@ -1,0 +1,1 @@
+export const PROGRESS_INDICATOR_TYPE_BAR = 'bar';
